Feat(all) -> Music, Tags, Author, Login, Category - Entities, Controllers and Repositories

+++ src/main/java/com/gmail/mateusfcosta2002/musicwebsite/Entities/converters/DurationConverter.java
@@ -6,14 +6,14 @@ import jakarta.persistence.Converter;
 import java.time.Duration;
 
 @Converter(autoApply = true)
-class DurationConverter implements AttributeConverter<Duration, Long> {
+class DurationConverter implements AttributeConverter<Duration, Integer> {
     @Override
-    public Long convertToDatabaseColumn(Duration attribute) {
-        return attribute.toSeconds();
+    public Integer convertToDatabaseColumn(Duration attribute) {
+        return Integer.valueOf((int)attribute.toSeconds());
     }
 
     @Override
-    public Duration convertToEntityAttribute(Long dbData) {
+    public Duration convertToEntityAttribute(Integer dbData) {
         return Duration.ofSeconds(dbData);
     }
 }
